# bindings/openapi.yaml
openapi: "3.0.0"
info:
  version: "1.0.0"
  title: Bindings for Pulp Ansible API
  license:
    name: Apache-2.0
paths:
...
  "/{prefix}/v3/collections/{namespace}/{name}/versions/":
    parameters:
      - in: path
        name: prefix
        required: true
        schema:
          type: string
      - in: path
        name: namespace
        required: true
        schema:
          type: string
      - in: path
        name: name
        required: true
        schema:
          type: string
    get:
      operationId: "list_versions"
      tags: ["galaxy: collections"]
      responses:
        200:
          description: 'OK'
          content:
            application/json:
              schema:
                type: object
  "/{prefix}/v3/collections/{namespace}/{name}/versions/{version}/":
    parameters:
      - in: path
        name: prefix
        required: true
        schema:
          type: string
      - in: path
        name: namespace
        required: true
        schema:
          type: string
      - in: path
        name: name
        required: true
        schema:
          type: string
      - in: path
        name: version
        required: true
        schema:
          type: string
    get:
      operationId: "get_version"
      tags: ["galaxy: collections"]
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic